[SWIG] moved CMake options for SWIG to src/CMakeLists.txt
When building mex files on Linux, all files need to be built with -fPIC. This
means that we need to find out in the top level if we're going to use matlab,
and if so, add appropriate flags for all compilations.
Sadly, this means that the top-level CMakeLists.txt now needs to know which
libraries we want to build with swig, breaking some modularity in the CMake files.