[SWIG] rename get_scanner_ptr() and get_proj_data_info_ptr() and make them safe
The C++ implementation of these 2 member functions has a vulnerability that it
can lead to a "dangling" pointer (if the original object is deleted). This needs
to be fixed in C++, but we currently only make the SWIG interface safe by returning
a const reference (which will be wrapped appropriatelly).
[Thanks to Vladimir Zeitlin for pointing this out.]

The functions are therefore renamed to get_scanner() and get_proj_data_info(),
which makes more sense in the target languages anyway.

WARNING: This breaks backwards compatibility for Python.