prevent OPENMP race conditions for accessing the ProjMatrixByBin cache
when using a lot of threads, there were conflicts between reading/writing
to the cache, resulting in locks or crashes. Now, these functions are
guarded with "omp critical" pragma. This seems to prevent errors
but slows it down dramatically. A better solution would be to have
explicit mutexes on a smaller scale (e.g. have a cache per view).