[OPENMP] use OPENMP 3.1 when available to minimise critical sections
We can use atomic-read etc to implement double-checked-locking in recent versions of OPENMP.
This avoids having to wait for all threads when accessing a cache (or singleton).

This commit uses this for the caches in ProjDataInfoCylindrical (and derived). To clean-up
the code, an additional private function is introduced where this checking occurs (as
opposed to have it wherever we use the cache).