[OPENMP] enhance parallelisation in scatter estimation
Use a dynamic scheduler for the scatter loop as computation time is
not predictable. This increases speed-up a lot in certain situations.

Also removed an old problem that we weree using static variables for the timing,
which would be problematic when having multiple scatter calculations in one
program. (However, only reported timings would have been wrong, results were correct).