[OPENMP] moved setting of number of threads to specific functions (changing the default as well)
Relevant functions are now in num_threads.h/.cxx files. They are also exposed via SWIG.
With the new set-up, the user can change the number of threads e.g. in Python/MATLAB
by using set_num_threads(). Previously, this had to be done by changing environment variables.

If OMP_NUM_THREADS is not set, we now default to using 90% of the number of processors
to avoid swamping your system.