Data type polymorphism - milestone 1.
A new base class for ProjData and CListModeData has been created. In
order to have a relation ProjData - ProjDataInfo the ExamInfo were
transfered in ExamData.

If possible, ExamData should be registered and ProjDataFromStream should
be a parsing object.

Right now this functionality has not been implemented. Therefore in
recontest.cxx the path to an input file has been hard coded.
Any ideas, if there is another way around it?

I set this as milestone 1 bacause is enough for what we need in the
scatter_correction branch.