More work on deriving data types from ExamData and seting additive and multiplicative sinograms in real-time.
The class MultipleProjData is derived from ExamData.

The GeneralisedObjectiveFunction hierarchy has functions to set the
additive and multiplicative sinograms.

The IterativeReconstruction has functions to set call the new set
function in the objective functions.

The AnalyticReconstruction is not going to be that straight-forward.
That's why I commit now before proceeding.