This commit addresses the following issues:
Still need to change get_projData_size to get_proj_data_size() (to
follow STIR naming convention)
:Done

If no changes in
PoissonLogLikelihoodWithLinearKineticModelAndDynamicProjectionData.txx,
can you revert it? It’d confuse people (and me).
: I replaced the code in the file with the code from the UCL/STIR:master
file.

In the test function, you still write the array to file. As far as I can
see, this is not really a test on the fill_from, copy_to functionality,
but just some writing. It creates a dependency on IO stuff which we
don’t need. In fact, this test would probably fail on Windows anyway
(need to open the stream binary). I’d cut it out.
:Done

Are we sure that the Reconstruction commit should be in this pull-
request? Maybe you should base your PR from something else? Could maybe
create a new one with a single clean commit (if it takes less than 10
mins).
: What I did was to sync my master with the upstream. Then I removed all
the leaked code manually (not that much) and merged with the current
branch.