[CMake] Changes to enable CMake import. WARNING: removed hand-crafted Makefile
we now create a stir/config.h (inccluded by stir/common.h) with preprocessor
defines etc as determined from CMake. This means they no longer need to be
passed via the command line (easier for external projects) and prevents having mixed
configurations.

However, this means that the old Makefile no longer works and had to be removed
(but it was quite obsolete anyway).