add tests for even span and introduce ProjDataInfo::construct_proj_data_info
Old terminology of ProjDataInfoCTI wasn't appropriate anymore. It is now
recommended to call construct_proj_data_info instead. It is identical
aside from the fact that it returns a std::unique_ptr<ProjDataInfo>
as opposed to a bare pointer.

Doxygen is clarified as well.

Added tests with span=2 and span=4.