Introduce even span (#103)
* When creating a ProjDataInfo, add support for GE-type axial compression, encoded as span=2
* Generalise this to other even span
* introduce ProjDataInfo::construct_proj_data_info as alternative to ProjDataInfo::ProjDataInfoCTI as its name is obsolete. The new version returns a unique_ptr.

Old terminology of ProjDataInfoCTI wasn't appropriate anymore. It is now
recommended to call construct_proj_data_info instead. It is identical
aside from the fact that it returns a std::unique_ptr<ProjDataInfo>
as opposed to a bare pointer.

Doxygen is clarified as well.

Added tests with span=2 and span=4.

* updated documentation and glossary for even span

* Use assignment with unique_ptr as opposed to reset

* More checks on span etc when constructing a ProjDataInfo

* use stir/unique_ptr.h

* corrected inadvertent change in batchmode