Bug fixes for TOF
This commit fixes some bugs found with Elise Emond and Nikos Efthimiou:

- Fixed bug in get_tof_bin (a delta in the last bin was assigned to bin
0). Also, let it return 0 if non-tof such that it is always usable.
- Fixed bugs in conversion from mm to TOF-time and vice versa (factor
2). This also needed a fix in the ROOT support (as the same factor 2 was
present).
- Removed "minimum timing step (in picoseconds)" keyword from
ROOT header as this should be hard-wired.
- introduced mm_to_tof_delta_time, tof_delta_time_to_mm and
get_tof_delta_time in ProjDataInfo
- fixed (probably irrelevant) bug in get_bin_for_det_pos_pair for
non-TOF data, where timing_pos was set undefined (and an assert was
thrown)
- let DetectionPositionPair constructors set timing_pos to 0 if not
specified as this is what non-TOF would have expected (this fixes an
assert in the test code).
- Add TOF-loops to test_proj_data_info

 Tests still fail on swapping of the TOF-bin.