safer setting of C++11 related choices
We were relying on run-time detection of std::shared_ptr and std::unique_ptr.
However, this created problems if STIR was built without C++11 but
imported somewhere else with C++11 (or the other way around).
This detection is now done by CMake. 2 options are provided to
switch back to boost::shared_ptr or std::auto_ptr for backwards compatibility.

The choices are not recorded in stir/config.h, such that any code that
uses installed version of STIR uses the correct choice.