More details on ScatterSimulation::subsample_image, and work on the test
The run_scatter_test was a bit weird as in the simultion sets the zoomed image both in for the scatter points and as the original attenuation image. I am not sure wether this is bug or a feature to speed up the process. I left it as is for now.