check consistency of set_up and actual projection/image data in projectors and BinNormalisation
We now check if set_up() has been called, and if it has been called
with the same type-of-data as what is being projected. This is really
the only way to be sure.

WARNING: This change requires derived classes to call set_up()
of the base-class.

WARNING: we currently keep a shared_ptr of the image used for set_up.
This means that if this shared_ptr gets changed, all tests
wll fail (TODO). However, it avoids memory overhead.