Handle 'maximum absolute segment to process' safer.
In the previous version, the projectors were set-up with a proj_data_info with
reduced number of segments. This was surprising and unsafe if the projectors in this
class were shared with someone else (as would happen e.g. in Python). Therefore,
the projectors are now set-up with the original data-size, as would be expected.

Note that this has a potential performance penalty for projectors who would
pre-store/calculate things. Overhead for current projectors is very small though.

Sharing of projectors is still dangerous if different data-sizes are used
for set-up though.

This partially resolves #51.