fixes #106
Introduced virtual InputStreamFromROOTFile::set_up(const std::string&). The function takes over functionality previously in post_processing(). But in addition the full path of the header file location can be used. 

FilePath is used for the merging of the strings. As this is the first time the class is used small modifications had to be made to make the functionality relevant. 

1. new function get_as_string() const; which returns the current my_string, wether it is path or file. 
2. If my_string has just a filename, function get_path_only) returns an empty string.