Read ParametricDiscretisedDensity
Beforehand, we used to read `ParametricDiscretisedDensity` as a `DynamicDiscretisedDensity`. This lead to problems, as `ParametricDiscretisedDensity`'s writer created the header with `num_time_frames` equal to 1. This meant that we were reading a `DynamicDiscretisedDensity` that only consisted of 1 time frame and therefore a `ParametricDiscretisedDensity` that only consisted of 1 kinetic parameter.

We needed to create a specific reader for `ParametricDiscretisedDensity` (a bit of copying from the `DynamicDiscretisedDensity` and then the method that converted the `DynamicDiscretisedDensity` to `ParametricDiscretisedDensity`).

We also needed some new keywords in the header, namely: `number of image data types` (number of kinetic parameters), `index nesting level` and `image data type description` (`:=` `slope` or `intercept`, for example).