IO updates (#208)
* Create a new file type `Multi` and enable dynamic projdata and densities to read it

* Read dynamic and parametric interfile format

* Write dynamic and parametric interfiles

* improved testing

* Corrections to IO updates

* Codacy improvements

* copyright additions, include fstream for test, avoid memory leak in read_from_file in DynamicProjData

* Set the exam info for a multipleProjData based on the first ProjData in the vector

* Error corrected in test_IO.h (check for null_ptr)

* fix parametricdiscretiseddensity IO test

* Add IO library to modelling_buildblock to be able to read ParametricDiscretisedDensity from file

* Read ParametricDiscretisedDensity

Beforehand, we used to read `ParametricDiscretisedDensity` as a `DynamicDiscretisedDensity`. This lead to problems, as `ParametricDiscretisedDensity`'s writer created the header with `num_time_frames` equal to 1. This meant that we were reading a `DynamicDiscretisedDensity` that only consisted of 1 time frame and therefore a `ParametricDiscretisedDensity` that only consisted of 1 kinetic parameter.

We needed to create a specific reader for `ParametricDiscretisedDensity` (a bit of copying from the `DynamicDiscretisedDensity` and then the method that converted the `DynamicDiscretisedDensity` to `ParametricDiscretisedDensity`).

We also needed some new keywords in the header, namely: `number of image data types` (number of kinetic parameters), `index nesting level` and `image data type description` (`:=` `slope` or `intercept`, for example).

* cant use nullptr as its c++11

* Run ctests with verbose output for debugging

* correct the new keys

* remove the `ctest --verbose` now that debugging is complete

* Moved `num_image_data_types` from `InterfileHeader` to `InterfileImageHeader`

Updated `documentation/release_3.1.htm`
Moved IO tests from `\ingroup test` `\ingroup IO`
Created virtual method, `get_num_data_types()`. Would be the product of time frames, image data types (for parametric images), gates, etc. Currently only the first two are implemented, and they are only implemented individually. That is to say, that at the moment, you can not have multiple time frames and data types at the same time.

* Added documentation for multi file format

* The second of the pair should be `end` not `duration`

* Created writer and reader for multi filetype

Included example format in `src/test/input/test_MultiOutputFileFormat.in`
Improved IO testing (for all filetypes)

* Update copyright and authors

* set_time_frame should be treated as if it is 1-based

* Changes in response to Nikos' comments. set_density instead of set_density_sptr. Should first set time_frame_definitions before setting densities with DynamicDiscretisedDensity (checks added to ensure this). Creating new VoxelsOnCartesianGrid will do deep copy of ExamInfo (before, pointer was shared and this could have strange effects if you change the original ExamInfo). Method create_shared_clone created for ExamInfo.

* remove shared_ptr from set_scanner. get_ methods const. use .at() instead of [] in TimeFrameDefinitions

* copyright and author changes

* DynamicProjData and MultipleProjData return unique_ptr instead of shared_ptr

* use boost as to_string is c++11

* Update release_3.1.htm

minor reordering