Revert flip_axes
After discussion with Kris, nifti images are always RAS (the voxel order isn't necessarily, but the coordinate system is). This means that the direction of the displacement field values are always RAS.

The ITK orienter will switch the voxel order, but does not touch the coordinate system, so it will still be in RAS after that. 

Therefore, flipping the direction of the displacement field components based on the ITK orienter doesn't make any sense. Remove it and always apply the same flips.

This will only be true for nifti images, so if not throw an error.