Prepend stir include files
We had a problem that ITK (and possibly other packages) added certain directories to the beginning of the list. On the cluster we had ITK and an older version of STIR in the same location. We were therefore including this older version of STIR, too, causing problems. 

We first do all the find_packages and then prepend our headers.