target_include_directories for boost
This avoids having to use find_package(boost), which should be avoided as much as possible. 

The ITK files were already linked privately, and so the find_package(ITK) can be removed from the STIRConfig.cmake file (which is good because the findITK.cmake prepends all its paths, which is a little naughty).