change checks for std smart pointers (#316)
Replace existing check for `std::shared_ptr` and `std::unique_ptr` that relied on Boost config with tests using `check_cxx_source_compile`. The tests are currently too simple, but are probably enough for the features that we need.

The main reason we needed this change is that Boost claims that Apple Clang's `std::unique_ptr` is not compliant. That might be true, but we don't see to have any problems using it (recently).

Closes #315