Initial commit
Process filesignature.

There is a weird point, as in order to get to the version I need to skip to byte 32 and read it as big endian. The header kind and header size are read correctly, as little endians in the right place.  In the print_header I cannot any relevant process.

Linking to SimSET  and Several initialisations working

Progress on matters of initialisations using SimSET headers.
Next step, initialise the history header and get events.

Fixed issues with the extern global variables in SimSET

Trying to guess the scanner.

CListModeSimSET tries to guess the scanner.

Ignore energy resolution, if has not been set in Scanner template