fix BUG in axial sampling for SPECT projection data
There was a left-over (presumably) of setting the "ring_spacing" to the
axial spacing * 2 from when we tried to fake SPECT data as "span=3" PET data.
(This was an attempt to let the reconstruction default to using the same number of slices
and spacing as the projection data. That strategy didn't work for an even number
of sinograms.)

Result was that STIR reported the axial sampling as twice too large when reading,
and therefore when copying the data, writing a file with twice the z-spacing.

This addresses #323.