Major Update - General RDP functionality incorporated into STIR and MATH
Added RDP to recon_buildblock_registries.

Parser Keys for RDP initialise keymap.

Replaced the quadratic penalty math with the RDP equations found :  J. Nuyts, D. Bequ, P. Dupont, and L. Mortelmans, “A Concave Prior Penalizing Relative Differences for Maximum-a-Posteriori Reconstruction in Emission Tomography,” vol. 49, no. 1, pp. 56–60, 2002. Eq. 5 and 6.

Introduced a new float (gamma) for the RDP - defaults to 1 but can be set by the parser.

Two functions, one to set the value of gamma, the other to return it.