Modification to the mathematics of the RDP value calculation.
In the RDP, each voxel pair is considered twice.  In the literature, when the derivative is calculated, this is not accounted. We take this into account by dividing the RDP value by two. 

This has been verified by comparing the finite difference, away from the limit zero, to the gradient.