Test class for Scatter simulation
Currently tests downsampling of the scanner and the image for scatter points. 
This commit might break the Scatter Estimation, as I made some changes in the input args of some functions. I made the changes having in mind the test flow and usability at SIRF. I will fix everything when re-working on the Scatter Estimation.