fix bug related to keyword order in Interfile image headers
`read_frames_info` did not use the total number of datasets when resizing, but just the number of time frames.
Therefore, if the `number of time frames` keyword occured after the `number of image data types`, the size was incorrect.

Also changed handling of `index_nesting_level` a bit. It still doesn't support multiple indices, but the previous code was misleading.

Fixes #346