removed unused ForwardProjectorByBin::forward_project(Bin&,...)
The implementation in terms of actual_forward_project got lost
in the previous merge, but it would need to be rewritten in terms
of the new style (not passing the image) anyway. In any case, most implementations
were calling error anyway.

As it is currently not used, it seemed easiest to just comment it out...