ScatterEstimation: correct a bug in 3D output and simplifed normalisatoin handling
In the previous version, the 3D scatter was "normalised", which was
probably unexpectedi by the user, but fine. However, the 3D additive
sinogram was computed by adding the "normalised" scatter with the
(unnormalised) randoms (and then normalising it with both normalisation
and attenuation factors). This was therefore incorrect!

In the current version, the 3D scatter output was "unnormalised" (i.e.
as the original data), and the 3D additive sinogram is (hopefully) correct.