don't estimate scatter scale factors when not necessary
upsample_and_fit_scatter_estimate computed scale factors even when the
thresholds are equal (and therefore any computation would be thrown away).
Now we don't do this anymore. This fix will therefore also reduce the likelihood
of #201 occurring, but it does not remove that problem.

Also cleaned-up the output of the scale_factors by putting them all in
`info` calls (with verbosity 2).