Change treatment of vectorised keys in KeyParser (backwards incompatible)
KeyParser did not store if a key is vectorised or not, but just assumed
that the stream it parsed was correct. This led to segmentation faults when
the stream used a vectorised key when KeyParser did not expect it, or vice versa.

This commit forces the developer to explicitly say what the level of vectorisation
is for a certain keyword (i.e. key, key[1], or key[1][1]).

This means that the change is backwards incompatible.

Classes using KeyParser have been rewritten to use the "shorter" form
of `add_key`, such that the vectorisation can be found by type-matching.

Also introduced a new `ignore_key()` member to make things more concise.

Fixes #148