add axpby for ProjData (#557)
Add `axpby` for `ProjData` (a*x+b*y). The generic version is quite slow as it works via `get/set_segment`. Specialisation for `ProjDataInMemory` is fast though as it works on the buffer.

`ProjDataInMemory` buffer is now directly a `float` array to enable this. This will also be useful for other things later.

Co-authored-by: Kris Thielemans <KrisThielemans@users.noreply.github.com>