fix inappropriate use of -DSWIG
To get the swig-generated wrappers to compile, we use #ifdef SWIG' in some .h files
to change some definitions. However, by using `add_definitions`, we set this
preprocessor variable also for compiling the registries.
This can fail if these rely on the normal STIR definitions.

This changes to use `-DSWIG` only for the wrappers.

Fixes #604