change Input registry for deformation fields
It was in terms of VoxelsOnCartesianGrid<CartesianCoordinate3D<float>>, but this was
incompatible with the change in read_from_file. I have switched it over to use
DiscretisedDensity<3,CartesianCoordinate3D<float>>. This is more future proof
anyway.

This needed changes in the ITK reader. Other code should not be affected.