Objective function setup - ignore the sensitivity recomputation
Set the sensitivity image to the filename used to load current_image_estimate_sptr/template_image_sptr and turn of recompute sensitivity. The sensitivity is not used in any of the kappa computation but we require objective function set_up() to be called for projectors. This method reduces the computation overhead as the sensitivity does not need recomputing