require ScatterEstimation::set_up() to be called
Formerly, process_data() called set_up(), but as that was virtual, that could lead
to surprising results as we don't know what a derived class needs. Therefore,
process_data() no longer does this, and requires the user to do it (as in other
places in STIR).

WARNING: this is a backwards incompatible change.