revise SinglesRates hierarchy
The SinglesRates classes assumed that stored singles were "rates" but in practice
were totals-per-time-interval (for both ECAT7 and RDF9). It is convenient
to be able to access totals and rates, so `SinglesRates` has now
`get_singles` (returning the singles in the time interval) and
`get_singles_rates` (returning the rate per second).

Due to the previous confusion, "client"-code will need to be adapted to use
either `get_singles` or `get_singles_rate`, depending on what is needed.