derive ProjDataInfoGeneric from ProjDataInfoCylindrical
Most of the code in Generic case in Cylindrical was the same. The Generic
case assumed a 2D system anyway. So I've cut out a lot of the duplication
by deriving Generical from Cylindrical. In fact, it should be the other way around,
but that requires more changes elsewhere. This is left for later.
At present, I just "delete" some member functions in Generic to make clear that
they shouldn't be called (they still can be of course, so this isn't very safe).

Added one noew function axial_sampling_is_uniform() to ProjDataInfoCylindrical
(and hence all Generic and Block versions). At present, this is still set to true.
It is used by the DataSymmetries to disable z-shift symmetry.