remove threshold in division by sensitivity [ci skip]
We used a small threshold to detect 0/0 when dividing by the
sensitivity image in the Poisson log likelihood. However, this threshold
was problematic, and actually caused problems in parametric imaging
due to different scales of the parametric images.

We now set that threshold to zero.

Addresses https://github.com/UCL/STIR/issues/906