TOF erf interpolation class (FastErf)  (#24)
This merge adds the FastErf class, and uses it in ProjMatrixByBin for TOF. FastErf is implemented to accelerate the computation of the y=erf(xp) function, which was found to be a significant portion of the TOF computation, by precomputing a vector of values and interpolating an approximate y for the given xp.

Three interpolations are added but only the linear interpolator is used in ProjMatrixByBin as a balance between computation speed and accuracy. See the PR (https://github.com/NikEfth/STIR/pull/24) for more details.