[CMake] fix find_package(HDF5) in exported CMake files
When HDF5 support is enabled, STIRConfig.cmake was
searching for the C libraries of HDF5, while STIR uses only
the C++ libraries. This was done in an attempt to work-around
a CMake problem that transpired in SIRF. However, the
work-around failed on some systems.

For pre CMake 3.21 we now first quitely search for both C and
CXX components of HDF5, before requiring the CXX component.
The first search will help some people who import STIR
and HDF5 C, although if the first search for HDF c, and then for
STIR, older CMake will still create a problem.