[SWIG] split stir.i and change order of declarations
The GeneralisedObjectiveFunction hierarchy was defined before
normalisation and projectors, leading to problems setting/getting
corresponding members.

As this was very messy, I've split stir.i in several files.
It still leads to one huge *wrap.cxx but at least reading stir.i
is now a bit easier. (still work to do).

This commit also fixes a line in SWIG for ProjData::get_subset,
but that function is still ignored, so it doesn't change the actual
Python interface.