fix openmp problem in ProjDataInfoCylindrical (#1088)
The change in #1084 was not thread-safe as the
ring_diff_arrays_computed variable was set before
all arrays were initialised. This meant that another
thread could believe this they were initialised anyway.

To make this change, I had to remove some assert()
in compute_segment_axial_pos_to_ring_pair that checked if
the function is the inverse of get_segment_axial_pos_num_for_ring_pair,
as this led to infinite recursion. However, that is checked
in test_proj_data_info anyway.

Fixes #1086