BinNormalisationFromProjdata::is_trivial now always returns false
We were going through all of the projection data to check if they
are all 1, but this takes time (especially when they are all 1).
In fact, it slowed us down in distributable_computation as it was
checked for every RelatedViewgrams.

As the case where the whole ProjData is set to 1 will never occur
in "real life", we save ourselves some time/complications
by returning false.