rewrite ProjDataInMemory to avoid streams
It was derived from ProjDataFromStream, which meant we had to write less code.
However, this meant it was slower than just doing the copies from memory directly.

Also, due to a limitation in boost::interprocess::bufferstream, total size
was limited to 2 GB.

Bits of the code are currently ugly to be compatible with both master and the TOF branch.

Fixes #1240