removed guessing code based on very old scanners
When parsing Interfile headers for projection data and the <tt>originating system</tt>
is not recognised, the previous version of STIR tried to guess the scanner based on the
number of views or rings. This was using very old scanners though, and could lead to
confusion. These guesses have now been removed.