enable branching ratio in BinNormalisatoonWithCalibration
we were not using the radionuclide in the exam_info used for set_up, but
an (uninitialised) class radionuclide class member, therefore always defaulting
to 1 (afer writing a warning).

However, this means that images will now have a different scale.

Also remove the get/set_branching_ratio members, as they shouldn't really
be part of the normalisation object.

This therefore breaks backwards-compatibility!