ProjData Visualisation window improvement and segment num bug fix (#1321)
- Fixes a bug with the segment_num parameter not being used in the method,
instead it was using an internal call to self.get_current_segment_num().
This limited the method to the current state of the data, rather than the abstract.

- Expand MLP canvas to full width

- Fix bug when a projdata file cannot be loaded (e.g. missing binary).
Previously, the projection data would be updated and fail when loading
the segment_data into memory. However, the stir_interface.projdata was updated.
This lead to errors with slider and reading new data. The PR uses temporary
variables when projdata and segment are loaded.

- Remove refresh_segment_data as this was leading to issues with the above
and recursive calls to load_projdata

- Add formatting to print_projdata_configuration and print_segment_data_configuration