Fast reconstruction for LAFOV scanners
* Building upon the cached lm reconstruction, we further parallelize the
calculation of the sensitity image to muliple passes when the scanner is
long.

* Now, a scanner is TOF ready even if it does not have set timing
resolution. If the timing resolution is 0 ps or if the TOF mashing leads
to a scanner with 1 TOF bin then we apply boundaries on the LOR's
length, equal to the coinc window. The probabilities do not change as we
set them to 1.

* cxx17 is nessesary for newer ROOT.