automatically use TOF sensitivity for TOF norm
We default use_tofsens=false, as this is the best for most scanners. However, for
Siemens Vision 600 (and others), the e7tools norm is TOF. This resulted in a
call to  error().

This update detects if the norm is BinNormalisationFromProjData with TOF data,
and sets use_tofsens=true if it is.