Auto use_tofsens (#1458)
automatically use TOF sensitivity calculation for TOF norm

We default use_tofsens=false, as this is the best for most scanners. However, for
Siemens Vision 600 (and others), the e7tools norm is TOF. This resulted in a
call to  error().

This update detects if the norm uses BinNormalisationFromProjData with TOF data,
and sets use_tofsens=true if it is. It does this by adding `is_TOF_only_norm()` to 
`BinNormalisation` (defaulting to `false`), and overriding it for certain classes.