Hessian times input: throw error if result would be incorrect
The Hessian calculations for the Poisson log-lik for projdata
use divide_and_truncate, which sets negative
numerators to zero. This is incorrect when the forward projection
of the "input" contains negatives.
We now check this and throw an error if it occurs. A proper fix
will have to be for later.

See https://github.com/UCL/STIR/issues/1461