add operator+ etc for Array, fix operator += in SWIG by ignoring
Explicitly add operator+, operator+= in Array. This overrides the
base-class implementations with functions that have the correct
return type.

This avoids surprising problems in python as well.

In addition, we work around https://github.com/swig/swig/issues/2634
by using %ignore all operator +=. This avoids a crash, but in fact
allows Python to fall back to operator+. This might be a bit slow,
but it does work.

Added testing both in C++ and Python.